+++ packages/create-plugin/src/messages.ts
@@ -31,7 +31,7 @@ const messages = {
     ja: "プラグインの日本語名を入力してください [1-64文字] (省略可)",
   },
   Q_NameJaError: {
-    en: "Plugin name must be within 64chars",
+    en: "Plug-in name must be within 64chars",
     ja: "プラグイン名は64文字以内で入力してください",
   },
   Q_DescriptionJa: {
@@ -39,8 +39,8 @@ const messages = {
     ja: "プラグインの日本語の説明を入力してください [1-200文字] (省略可)",
   },
   Q_DescriptionJaError: {
-    en: "Plug-in description must be within 64chars",
-    ja: "プラグインの日本語の説明を64文字以内で入力してください",
+    en: "Plug-in description must be within 200chars",
+    ja: "プラグインの日本語の説明を200文字以内で入力してください",
   },
   Q_SupportZh: {
     en: "Does your plug-in support Chinese?",
@@ -59,8 +59,28 @@ const messages = {
     ja: "プラグインの中国語の説明を入力してください [1-200文字] (省略可)",
   },
   Q_DescriptionZhError: {
-    en: "Plug-in description must be within 64chars",
-    ja: "プラグインの中国語の説明を64文字以内で入力してください",
+    en: "Plug-in description must be within 200chars",
+    ja: "プラグインの中国語の説明を200文字以内で入力してください",
+  },
+  Q_SupportEs: {
+    en: "Does your plug-in support Spanish?",
+    ja: "スペイン語をサポートしますか？",
+  },
+  Q_NameEs: {
+    en: "Input your plug-in name in Spanish [1-64chars] (Optional)",
+    ja: "プラグインのスペイン語名を入力してください [1-64文字] (省略可)",
+  },
+  Q_NameEsError: {
+    en: "Plug-in name must be within 64chars",
+    ja: "プラグイン名は64文字以内で入力してください",
+  },
+  Q_DescriptionEs: {
+    en: "Input your plug-in description in Spanish [1-200chars] (Optional)",
+    ja: "プラグインのスペイン語の説明を入力してください [1-200文字] (省略可)",
+  },
+  Q_DescriptionEsError: {
+    en: "Plug-in description must be within 200chars",
+    ja: "プラグインのスペイン語の説明を200文字以内で入力してください",
   },
   Q_WebsiteUrlEn: {
     en: "Input your home page url for English (Optional)",
@@ -73,6 +93,10 @@ const messages = {
   Q_WebsiteUrlZh: {
     en: "Input your home page url for Chinese (Optional)",
     ja: "プラグインの中国語のWebサイトURLを入力してください (省略可)",
+  },
+  Q_WebsiteUrlEs: {
+    en: "Input your home page url for Spanish (Optional)",
+    ja: "プラグインのスペイン語のWebサイトURLを入力してください (省略可)",
   },
   Q_MobileSupport: {
     en: "Does your plug-in support mobile views?",